*) Code from Purga MISC NOTES - Changed: Reactivated the Icon field in the Style database, this is used instead of ID as the Style 1hed animation. - Added: BlankBrain can be used when you want to add a GameNPC but don't want a brain. - Bugfix: Guard, Protect and Intercept's event handlers for group disband is now correctly cleaned up this prevents multiple you are no longer protecting xxx messages. - Added: The effect description (the mouseover text of the effect icons) now display the guard target and the guard source - Changed: Housing a little, it still isn't ready - Bugfix: There are now 10% TOA buffbonus caps - New: You can now view your deathblow and solo kill stats in the title window - Bugfix: You can no longer use Stag or Dirty Tricks when Mezzed, Stunned or Sitting - New: Formula for PA/BS - Changed: You can now gain a percentage of experience and realm points while dead - Changed: Vampiir class will now remove the spells and abilitys he shouldnt have - Bugfix: You can now recharge and repair items that you can't drop - New: CamouflageEffect (Does nothing until Mastery of Stealth is added) - New: PvE Missrate now takes into account the group leaders level - New: Support for multiple targets with Savage Styles - Changed: GMs ability timers are no longer disabled, GMs can now detect all stealthed players regardless of range, GMs stealth cannot be detected SCRIPTS - Changed: /code can now make use of the namespace System.Collections and DOL.GS.Movement - Bugfix: When you change someones plvl, a message will output telling you about it - Bugfix: Bonedancers and Warlocks can now become Tailors - New: /mob questinfo command, to see what quests a mob can give - Changed: The temporary graphic for the pathpoints, from a bag to a groundtarget - New: Emotes and commands, /sweat, /stagger, /yawn, /doh, /confuse, /shiver, /rofl, /mememe, /worship, /drink - New: /who will now display CG, GM and ADMIN as valid SPELLS - Changed: Stealth spec is now capped at 50 effectiveness - Changed: Damage adds and shields now have a DPS cap - Changed: DOT Spells now are affected by your mana stat - Changed: Lifedrain spells are now affected by your spec level - Changed: Lifetransfer spells are no longer affected by your maxhealth - Changed: Various proc spells should be overwriteable now - New: Spellhandlers for Thrust and Crush resist debuffs - New: Spellhandlers for DPS, Mez dampening, Evade, Parry, WeaponSkill, StealthSkill, ToHit and MagicResists buffs - New: Spellhandlers for DPS and Skills debuffs - New: Formula for Dex affecting cast time - New: Formula for spell damage CRAFTING - Bugfix: Attempting to craft while you are crafting will not only stop the craft timer, but close the craft timer window. - Changed: When crafting the chance to make an item and chance to gain a skill will display - Changed: Base chance of success for the last 3 tiers of crafting from 40,35,35 to 45,45,45 - Changed: When repairing the chance of repairing an item will display - Added: Code to support the repairing of siege weapons - Bugfix: Failing to build a spellcraft item will no longer result in 100% chance for dying - Bugfix: You no longer have to be at a lathe to craft a siege weapon GAMEOBJECTS - New: Rewritten siege weapons - New: Mobs with ranged weapons should now be able to use them intelligently - Bugfix: Attempting to use a charge when it is above your level will no longer use up the charge - New: Block and evade chance is now affected by shield size and attack count - New: If you are fully healed all attackers and xp gainers are cleared from you - BugFix: All versions of spreadheals are now accessible once you have learnt them - Changed: New Realm Point Formula - Changed: You can only miss your attacks when strafing against players - Added: Messages to tell you when your items should need repairs - Changed: Styles are now displayed in lists like spells - Bugfix: Gaining realm points will advance your realm rank by the amount appropriate - New: NPC Templates now load styles - Bugfix: When bladeturn is hit and you are stealthed, you are unstealthed PACKETS - Changed: CheckLOSRequest now accepts targetOID as an argument, since you can check LOS for others - Bugfix: You can no longer destroy items currently equipped - Bugfix: Left click on portal keeps and other keep doors should behave as it should now - New: You can now detect clients logging in with packet loggers *) Code from Me - Changed: When trying to /send to a player that is loading will tell you so, instead of saying they can't be found. Quote: - New: code to allow most messages to be sent while a player is loading Quote: - New: /report command to report bugs and BugReport table to view the bugs - New: The database will now automatically convert the mobs and styles to the format needed - Bugfix: Mobs with realm < realm first and > realm last will no longer error the server with an index out of range exception in the npcguildscript *) Code From Blue Fields are now automatically created if they are missing *) Code from Purga based on work by Avithan - New: Support for BP Merchant, Seal and Aurulite merchants.
git-svn-id: https://svn.code.sf.net/p/dolserver/code/DOLSharp/branch/1.x/GameServer@150 c9877f0d-d2ea-46ae-93cb-121acdf3e971

+++ housing/IndoorItem.cs
@@ -41,14 +41,14 @@ namespace DOL.GS.Housing
 			get { return m_color; }
 			set { m_color = value;}
 		}
-		int m_x;
-		public int X
+		short m_x;
+		public short X
 		{
 			get { return m_x; }
 			set { m_x = value;}
 		}
-		int m_y;
-		public int Y
+		short m_y;
+		public short Y
 		{
 			get { return m_y; }
 			set { m_y = value;}
@@ -91,14 +91,14 @@ namespace DOL.GS.Housing
 		}
 		public void CopyFrom(DBHouseIndoorItem dbitem)
 		{
-	    	this.Model = dbitem.Model;
-	    	this.Color = dbitem.Color;
-	    	this.X = dbitem.X;
-	    	this.Y = dbitem.Y;
-	    	this.Rotation = dbitem.Rotation;
-	    	this.Size = dbitem.Size;
-	    	this.Position = dbitem.Position;
-	    	this.Placemode = dbitem.Placemode;
+			this.Model = dbitem.Model;
+			this.Color = dbitem.Color;
+			this.X = (short)dbitem.X;
+			this.Y = (short)dbitem.Y;
+			this.Rotation = dbitem.Rotation;
+			this.Size = dbitem.Size;
+			this.Position = dbitem.Position;
+			this.Placemode = dbitem.Placemode;
 			this.BaseItem = (ItemTemplate)GameServer.Database.FindObjectByKey(typeof(ItemTemplate), dbitem.BaseItemID);
 			this.DatabaseItem = dbitem;
 		}
